Added Un'Goro and fixed numerous bugs.
There are now two save files: the card collection and a quantity
map. This means that any further update to the card collection format
will NOT break your collection. On top of this, save files are now
automatically migrated on update. The quantity of a card is now a Maybe
Quantity, to represent the absence of information for such a card.

The in-memory representation is now a map from id to full card; this
change now allows slightly more complex filtering. This has been used to
fix the "missing" filter, which used to match Legendaries of quantity 1.

Filters now support a pseudo-OR; it is possible to do r=Epic,Legendary
or h=Druid,Warrior for instance. There is also a "standard" filter.

Cleaned error and info messages a bit.

The input command now saves after each successful input.

Added a fix command that simply prompts all new cards.